requirements.txt
# Core Django
Django>=4.2,<5.0
django-crispy-forms>=2.0
crispy-bootstrap4>=2.0

# Database
psycopg2-binary>=2.9  # PostgreSQL adapter for production

# Security & Authentication
django-ratelimit>=4.1.0  # Rate limiting for security

# Utilities
python-dotenv>=1.0.0  # Environment variable management
pillow>=10.0.0  # Image processing for user avatars

# Geolocation
geopy>=2.4.0  # Geocoding service

# Data Export
reportlab>=4.0.0  # PDF generation
xlsxwriter>=3.1.0  # Excel export

# API & Real-time (Optional - for Phase 4+)
# djangorestframework>=3.14.0
# channels>=4.0.0  # WebSocket support
# daphne>=4.0.0  # ASGI server

# SMS/Notifications (Optional - uncomment when needed)
# twilio>=8.10.0  # SMS notifications

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0

# Code Quality
flake8>=6.1.0
black>=23.7.0

# Production Server
gunicorn>=21.2.0  # WSGI HTTP Server
whitenoise>=6.5.0  # Static file serving

# Monitoring & Logging
sentry-sdk>=1.35.0  # Error tracking (optional)

# Caching (Optional - for production)
# redis>=5.0.0
# django-redis>=5.3.0

# Development Tools
django-debug-toolbar>=4.2.0  # Only for development
